+++ source/index.ts
@@ -1,0 +1,3 @@
+import Popover from "./components/popover"
+
+export default Popover
