+++ source/lib/platform.ts
