+++ stories/rows/main.tsx
@@ -1,6 +1,6 @@
 import F from "ramda"
 import React from "react"
-import Popover from "../../source"
+import Popover from "../../source/components/popover"
 import "../base.css"
 import "./main.css"
